+++ backend/src/product/product.module.ts
@@ -9,6 +9,7 @@ import { TrackingProductRepository } from './trackingProduct.repository';
 import { MongooseModule } from '@nestjs/mongoose';
 import { ProductPrice, ProductPriceSchema } from 'src/schema/product.schema';
 import { FirebaseService } from 'src/firebase/firebase.service';
+import { CronService } from 'src/cron/cron.service';
 
 @Module({
     imports: [
@@ -16,6 +17,6 @@ import { FirebaseService } from 'src/firebase/firebase.service';
         MongooseModule.forFeature([{ name: ProductPrice.name, schema: ProductPriceSchema }]),
     ],
     controllers: [ProductController],
-    providers: [ProductService, ProductRepository, TrackingProductRepository, FirebaseService],
+    providers: [ProductService, ProductRepository, TrackingProductRepository, FirebaseService, CronService],
 })
 export class ProductModule {}
