+++ backend/src/product/product.service.ts
@@ -13,13 +13,9 @@ import { ProductPrice } from 'src/schema/product.schema';
 import { Model } from 'mongoose';
 import { ProductPriceDto } from 'src/dto/product.price.dto';
 import { PriceDataDto } from 'src/dto/price.data.dto';
-import { CHANNEL_ID, MAX_TRACKING_RANK, NINETY_DAYS, NO_CACHE, THIRTY_DAYS } from 'src/constants';
-import { Cron } from '@nestjs/schedule';
+import { MAX_TRACKING_RANK, NINETY_DAYS, NO_CACHE, THIRTY_DAYS } from 'src/constants';
 import { ProductRankCache } from 'src/utils/cache';
 import { ProductRankCacheDto } from 'src/dto/product.rank.cache.dto';
-import { FirebaseService } from '../firebase/firebase.service';
-import { Message } from 'firebase-admin/lib/messaging/messaging-api';
-import { TrackingProduct } from 'src/entities/trackingProduct.entity';
 import Redis from 'ioredis';
 import { InjectRedis } from '@songkeys/nestjs-redis';
 
@@ -36,7 +32,6 @@ export class ProductService {
         @InjectModel(ProductPrice.name)
         private productPriceModel: Model<ProductPrice>,
         @InjectRedis() private readonly redis: Redis,
-        private readonly firebaseService: FirebaseService,
     ) {
         this.initCache();
     }
@@ -288,123 +283,6 @@ export class ProductService {
         });
     }
 
-    @Cron('* */10 * * * *')
-    async cyclicPriceChecker() {
-        const productList = await this.productRepository.find({ select: { id: true, productCode: true } });
-        const productCodeList = productList.map(({ productCode, id }) => getProductInfo11st(productCode, id));
-        const results = await Promise.all(productCodeList);
-        const updatedDataInfo: ProductInfoDto[] = [];
-        await Promise.all(
-            results.map(async (data) => {
-                const { productId, productPrice, isSoldOut } = data;
-                const cacheData = await this.redis.get(`product:${productId}`);
-                const cache = JSON.parse(cacheData as string);
-                if (!cache || cache.isSoldOut !== isSoldOut || cache.price !== productPrice) {
-                    const lowestPrice = cache ? Math.min(cache.lowestPrice, productPrice) : productPrice;
-                    await this.redis.set(
-                        `product:${productId}`,
-                        JSON.stringify({
-                            isSoldOut,
-                            price: productPrice,
-                            lowestPrice,
-                        }),
-                    );
-                    updatedDataInfo.push(data);
-                }
-            }),
-        );
-        if (updatedDataInfo.length > 0) {
-            await this.productPriceModel.insertMany(
-                updatedDataInfo.map(({ productId, productPrice, isSoldOut }) => {
-                    return { productId, price: productPrice, isSoldOut };
-                }),
-            );
-            const { messages, products } = await this.getNotifications(updatedDataInfo);
-            if (messages.length > 0) {
-                const { responses } = await this.firebaseService.getMessaging().sendEach(messages);
-                const successProducts = products.filter((item, index) => {
-                    const { success } = responses[index];
-                    item.isFirst = false;
-                    return success;
-                });
-                if (successProducts.length > 0) {
-                    await this.trackingProductRepository.save(successProducts);
-                }
-            }
-        }
-    }
-    getMessage(
-        productCode: string,
-        productName: string,
-        productPrice: number,
-        imageUrl: string,
-        token: string,
-    ): Message {
-        return {
-            notification: {
-                title: '목표 가격 이하로 내려갔습니다!',
-                body: `${productName}의 현재 가격은 ${productPrice}원 입니다.`,
-            },
-            data: {
-                productCode,
-            },
-            android: {
-                notification: {
-                    channelId: CHANNEL_ID,
-                    imageUrl,
-                },
-            },
-            token,
-        };
-    }
-    async getNotifications(
-        productInfo: ProductInfoDto[],
-    ): Promise<{ messages: Message[]; products: TrackingProduct[] }> {
-        const productIds = productInfo.map((p) => p.productId);
-
-        const trackingProducts = await this.trackingProductRepository
-            .createQueryBuilder('tracking_product')
-            .where('tracking_product.productId IN (:...productIds)', { productIds })
-            .getMany();
-
-        const trackingMap = new Map<string, TrackingProduct[]>();
-        trackingProducts.forEach((tracking) => {
-            const products = trackingMap.get(tracking.productId) || [];
-            trackingMap.set(tracking.productId, [...products, tracking]);
-        });
-        const results = await Promise.all(
-            productInfo.map(async ({ productId, productCode, productName, productPrice, imageUrl }) => {
-                const trackingList = productId ? trackingMap.get(productId) || [] : [];
-                const notifications = [];
-                const matchedProducts = [];
-
-                for (const trackingProduct of trackingList) {
-                    const { userId, targetPrice, isFirst, isAlert } = trackingProduct;
-                    if (!isFirst && targetPrice < productPrice) {
-                        trackingProduct.isFirst = true;
-                        await this.trackingProductRepository.save(trackingProduct);
-                    } else if (targetPrice >= productPrice && isFirst && isAlert) {
-                        const firebaseToken = await this.redis.get(`firebaseToken:${userId}`);
-                        if (firebaseToken) {
-                            notifications.push(
-                                this.getMessage(productCode, productName, productPrice, imageUrl, firebaseToken),
-                            );
-                            matchedProducts.push(trackingProduct);
-                        }
-                    }
-                }
-                return { notifications, matchedProducts };
-            }),
-        );
-
-        const allNotifications = results.flatMap((result) => result.notifications);
-        const allMatchedProducts = results.flatMap((result) => result.matchedProducts);
-
-        return {
-            messages: allNotifications,
-            products: allMatchedProducts,
-        };
-    }
     async firstAddProduct(productCode: string) {
         const productInfo = await getProductInfo11st(productCode);
         const product = await this.productRepository.saveProduct(productInfo);
